TSK-1415 unify date selector
