+++ packages/ui/src/components/calendar/DateRangePicker.svelte
@@ -24,7 +24,6 @@
   export let value: number | null | undefined = null
   export let withTime: boolean = false
   export let icon: 'normal' | 'warning' | 'overdue' = 'normal'
-  export let labelOver: IntlString | undefined = undefined
   export let labelNull: IntlString = ui.string.NoDate
 
   const dispatch = createEventDispatcher()
@@ -40,14 +39,6 @@
 <div class="antiSelect antiWrapper cursor-default">
   <div class="flex-col">
     <span class="label mb-1"><Label label={title} /></span>
-    <DateRangePresenter
-      {value}
-      mode={DateRangeMode.DATETIME}
-      {icon}
-      {labelOver}
-      {labelNull}
-      editable
-      on:change={changeValue}
-    />
+    <DateRangePresenter {value} mode={DateRangeMode.DATETIME} {icon} {labelNull} editable on:change={changeValue} />
   </div>
 </div>
