+++ packages/ui/src/index.ts
@@ -63,7 +63,6 @@ export { default as ColorPopup } from './components/ColorPopup.svelte'
 export { default as TextArea } from './components/TextArea.svelte'
 export { default as TextAreaEditor } from './components/TextAreaEditor.svelte'
 export { default as Section } from './components/Section.svelte'
-export { default as DatePickerPopup } from './components/calendar/DatePickerPopup.svelte'
 export { default as DatePicker } from './components/calendar/DatePicker.svelte'
 export { default as DateRangePicker } from './components/calendar/DateRangePicker.svelte'
 export { default as DatePopup } from './components/calendar/DatePopup.svelte'
